feat: moved qrcode controller to web.rest package, to be more inline with other rest endpoints
