+++ payment-qrlink/src/main/kotlin/nl/nlportal/payment/qrlink/web/rest/QRLinkPaymentController.kt
@@ -1,25 +1,8 @@
-/*
- * Copyright 2015-2023 Ritense BV, the Netherlands.
- *
- * Licensed under EUPL, Version 1.2 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nl.nlportal.payment.qrlink.api
+package nl.nlportal.payment.qrlink.web.rest
 
 import nl.nlportal.core.qrcode.QRCodeFileExtension
-import nl.nlportal.payment.qrlink.domain.QRLinkPaymentCodeFileType
 import nl.nlportal.payment.qrlink.domain.QRLinkPaymentResponse
 import nl.nlportal.payment.qrlink.service.QRLinkPaymentService
-import org.apache.commons.io.FileUtils
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
